% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anexoII}
\alias{anexoII}
\title{Tabela do Anexo II da Resolução CONSUNI 18/2017}
\format{
\subsection{'anexoII'}{

A tibble: 232 × 3
\describe{
\item{COD}{Contém o código da atividade.}
\item{Descrição}{Contém a descrição da atividade.}
\item{Pontos}{Contém a pontuação da atividade.}
}

@source https://sistemas.ufg.br/consultas_publicas/resolucoes/arquivos/Resolucao_CONSUNI_2017_0018.pdf
}
}
\usage{
anexoII
}
\description{
Contém a Tabela do Anexo II com suas respectivas pontuações
}
\keyword{datasets}
