% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad.R
\name{cad}
\alias{cad}
\title{Retorna o dataframe pronto para impressão da Tabela Cad}
\usage{
cad(df, n = 40)
}
\arguments{
\item{df}{dataframe com os dados do RADOC}
}
\value{
um dataframe com a Tabela CAD preenchida
}
\description{
Retorna o dataframe pronto para impressão da Tabela Cad
}
\examples{
\dontrun{
cad(df)
}
}
