% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cadOrig}
\alias{cadOrig}
\title{Tabela de Pontuação da CAD}
\format{
\subsection{'cadOrig'}{

A tibble: 24 × 2
\describe{
\item{Item}{Contém o Item da ser avaliado}
\item{Avaliação de desempenho para Progressão}{Contem  uma breve descrição do item a ser avaliado}
}

@source https://sistemas.ufg.br/consultas_publicas/resolucoes/arquivos/Resolucao_CONSUNI_2017_0018.pdf
}
}
\usage{
cadOrig
}
\description{
Contém a Tabela de pontuação da CAD, com os itens a serem pontuados, separados por grupos
}
\keyword{datasets}
