% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coleta.R
\name{coleta}
\alias{coleta}
\title{Extrai os dados do Docente}
\usage{
coleta(file)
}
\arguments{
\item{file}{arquivo em pdf do RADOC do docente}
}
\value{
um dataframe com os dados coletados no arquivo
}
\description{
Extrai os dados do Docente
}
\examples{
\dontrun{
coleta('data/radoc1.pdf')
}
}
