% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docente.R
\name{dados_docente}
\alias{dados_docente}
\title{Extrai os dados do Docente do arquivo PDF}
\usage{
dados_docente(pdf_file)
}
\arguments{
\item{pdf_file}{arquivo PDF com os dados do RADOC}
}
\value{
um dataframe com os dados do docente
}
\description{
Extrai os dados do Docente do arquivo PDF
}
\examples{
\dontrun{
dados_docente(pdf_file)
}
}
\keyword{internal}
