% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filesCheck}
\alias{filesCheck}
\title{Realiza verificações em arquivos RADOC}
\usage{
filesCheck(...)
}
\arguments{
\item{...}{Um ou mais caminhos para arquivos PDF contendo dados do RADOC.
Cada PDF deve conter RADOCs gerados pelo SICAD+.}
}
\value{
Uma mensagem indicando se os arquivos passaram nos testes.
}
\description{
A função \code{filesCheck()} faz três testes em um ou mais arquivos PDF de RADOC,
que devem ter sido gerados pelo SICAD+:
}
\details{
\enumerate{
\item Verifica se cada arquivo é um RADOC válido do SICAD+.
\item Verifica se os RADOCs fornecidos pertencem ao mesmo docente.
\item Verifica se os RADOCs fornecidos são de anos diferentes.
}
}
\examples{
\dontrun{
# Exemplo com dois arquivos RADOC
filesCheck("radoc1.pdf", "radoc2.pdf")
}
}
