% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_cad.R
\name{final_cad}
\alias{final_cad}
\title{umn dataframe com a Tabela da CAD dos dados}
\usage{
final_cad(pdf_file1, pdf_file2, n = 40)
}
\arguments{
\item{pdf_file1, }{arquivo PDF com os dados do 1o RADOC}

\item{pdf_file2, }{arquivo PDFcom os dados do 2o RADOC}

\item{n, }{carga horária do docende (padrão n=40)}
}
\value{
um dataframe com a Tabela CAD preenchida
}
\description{
umn dataframe com a Tabela da CAD dos dados
}
\examples{
\dontrun{
final_cad(pdf_file1, pdf_file2, n=(20 ou 40))
}
}
