% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_doc.R
\name{final_doc}
\alias{final_doc}
\title{docente}
\usage{
final_doc(pdf_file1, pdf_file2)
}
\arguments{
\item{pdf_file1, }{arquivo PDF com os dados do 1o RADOC}

\item{pdf_file2, }{arquivo PDFcom os dados do 2o RADOC}
}
\value{
um dataframe com os dados do docente
}
\description{
docente
}
\examples{
\dontrun{
final_doc(pdf_file1, pdf_file2)
}
}
