% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lista0}
\alias{lista0}
\title{Lista0}
\format{
\subsection{lista0}{

\describe{
\item{"II-1-20"}{}
\item{"II-3-9"}{}
\item{"II-3-19.2"}{}
\item{"..."}{}
}
}
}
\usage{
lista0
}
\description{
Contém a lista de tópicos pontuados considerando o tempo no cargo
}
\keyword{datasets}
