% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{listas}
\alias{listas}
\title{Listas}
\format{
\subsection{listas}{

\describe{
\item{"lista_tempo"}{lista com os itens pontuados por tempo da atividade}
\item{"lista1"}{lista com os itens limitados a 1 evento por ano}
\item{"lista3"}{lista com os itens limitados a 3 pontos por ano}
\item{"lista4"}{lista com os itens limitados a 4 pontos por ano}
\item{"lista9"}{lista com os itens limitados a 9 pontos por ano}
\item{"lista10"}{lista com os itens limitados a 10 pontos por ano}
\item{"lista12"}{lista com os itens limitados a 12 pontos por ano}
\item{"lista15"}{lista com os itens limitados a 15 pontos por ano}
\item{"lista20"}{lista com os itens limitados a 20 pontos por ano}
\item{"lista24"}{lista com os itens limitados a 24 pontos por ano}
\item{"lista30"}{lista com os itens limitados a 30 pontos por ano}
\item{"lista40"}{lista com os itens limitados a 40 pontos por ano}
\item{"lista_anual"}{lista com os itens pontuodos por cada ano de atividade}
\item{"lista150"}{lista com os itens limitados a 10 pontos a cada 150 horas}
}
}
}
\usage{
listas
}
\description{
Contém as listas de tópicos com pontuações diferenciadas
}
\keyword{datasets}
