% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtibble.R
\name{mtibble}
\alias{mtibble}
\title{Função que transforma um tibble em um tibble construction, i.e.,
a função retorna o modo de construção do tibble de entrada}
\usage{
mtibble(df)
}
\arguments{
\item{df}{Um dataframe do tipo tibble}
}
\value{
o modo de construção do tibble de entrada
}
\description{
Função que transforma um tibble em um tibble construction, i.e.,
a função retorna o modo de construção do tibble de entrada
}
\examples{
\dontrun{
mtibble(iris)
}
}
