% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_cad.R
\name{pontua}
\alias{pontua}
\title{Pontua os dados do RADOC seguindo o Anexo II da Resolução Consuni 18/2017}
\usage{
pontua(pdf_file)
}
\arguments{
\item{pdf_file}{arquivo em pdf RADOC gerado pelo SICAD+}
}
\value{
um dataframe com os dados e suas respectivas pontuações
}
\description{
Pontua os dados do RADOC seguindo o Anexo II da Resolução Consuni 18/2017
}
\examples{
\dontrun{
pontua(pdf_file)
}
}
