% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radocr-package.R
\docType{package}
\name{radocr-package}
\alias{radocr}
\alias{radocr-package}
\title{radocr: Este Pacote extrai os dados de um RADOC gerado pelo SICAD+}
\description{
Este pacote extrai os dados de um arquivo em pdf contendo o RADOC gerado pelo SICAD+ e apresenta um resumo dos dados.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
