% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumo.R
\name{resumo_tex}
\alias{resumo_tex}
\title{Gera um resumo em LaTeX das atividades docentes a partir de arquivos RADOC}
\usage{
resumo_tex(...)
}
\arguments{
\item{...}{Um ou mais caminhos para arquivos PDF contendo os RADOCs do
docente.}
}
\value{
Um arquivo \code{resumo.tex} que contém o código LaTeX do resumo das
atividades do docente.
}
\description{
A função \code{resumo_tex()} lê os dados de um ou mais arquivos PDF de RADOC e
gera um arquivo LaTeX (.tex) contendo um resumo das atividades docentes no
período. É útil para criar relatórios personalizados de atividades docentes.
}
\details{
Você pode fornecer quantos arquivos de RADOC quiser, e a função os combinará
em um único documento LaTeX. Os arquivos devem ser gerados pelo sistema SICAD+.
}
\examples{
\dontrun{
# Caminho para os arquivos RADOC
pdf_radoc1 <- "pasta_do_arquivo/radoc1.pdf"
pdf_radoc2 <- "pasta_do_arquivo/radoc2.pdf"

# Gera o arquivo LaTeX com o resumo
resumo_tex(pdf_radoc1, pdf_radoc2)
}

}
