% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumo.R
\name{resumo_xlsx}
\alias{resumo_xlsx}
\title{Acrescenta o resumo das atividades docentes na tabela criada pela função
tabela_cad}
\usage{
resumo_xlsx(wb_cad, pdf_files)
}
\arguments{
\item{wb_cad}{planilha criada pela função tabela_cad.}

\item{pdf_files}{o caminho para os arquivos contendo os RADOCs.}
}
\description{
Esta função lê os dados do RADOC em 2 arquivos e adiciona o resumo na
planilha criada pela função tabela_cad
}
\examples{
 \dontrun{
wb_cad <- "planilha com a tabela cad"
pdf_files <- "arquivos pdf com os radocs"
resumo_xlsx(wb_cad, pdf_files)
}
}
