% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_cad.R
\name{tabela_cad}
\alias{tabela_cad}
\title{Gera uma tabela CAD em formato Excel a partir de arquivos RADOC em PDF}
\usage{
tabela_cad(...)
}
\arguments{
\item{...}{Um ou mais caminhos para arquivos PDF contendo dados do RADOC.
Cada PDF deve conter RADOCs gerados pelo SICAD+.}
}
\value{
Um arquivo \code{tabela_cad.xlsx} contendo a tabela CAD gerada a partir dos PDFs fornecidos.
}
\description{
A função \code{tabela_cad()} gera uma tabela CAD em formato Excel usando
dados contidos em um ou mais arquivos PDF de RADOC.
Os PDFs devem conter RADOCs gerados pelo SICAD+.
}
\examples{
\dontrun{
# Exemplo com um arquivo RADOC
tabela_cad("radoc1.pdf")

# Exemplo com múltiplos arquivos RADOC
tabela_cad("radoc1.pdf", "radoc2.pdf")
}

}
