% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_cad.R
\name{tabela_cad_tex}
\alias{tabela_cad_tex}
\title{Cria a Tabela Cad (em LaTeX) a partir dos RADOCs}
\usage{
tabela_cad_tex(pdf_file1, pdf_file2, output_file = NULL)
}
\arguments{
\item{pdf_file1}{o caminho para o arquivo contendo o 1o RADOC.}

\item{pdf_file2}{o caminho para o arquivo contendo o 2o RADOC.}

\item{output_file}{o caminho onde o arquivo de saída deve ser salvo.}
}
\description{
Esta função lê os dados do RADOC em 2 arquivos PDF e retorna um
arquivo LaTeX (.tex) com a Tabela Cad preenchida.
}
\details{
Em caso de ausência do argumento \code{output_file} o arquivo será salvo na
pasta atual.
}
\examples{
\dontrun{
pdf_radoc1 <- "pasta_do_arquivo/radoc1.pdf"
pdf_radoc2 <- "pasta_do_arquivo/radoc2.pdf"
output_file <- "pasta_de_destino/nome_arquivo.tex"
tabela_cad_tex(pdf_radoc1, pdf_radoc2, output_file)
}
}
