% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_cad_xlsx.R
\name{tabela_cad_xlsx}
\alias{tabela_cad_xlsx}
\title{Create a Exccel Document from PDF Data}
\usage{
tabela_cad_xlsx(pdf_file1, pdf_file2, output_file, n = 40)
}
\arguments{
\item{pdf_file1}{Path to the PDF file 1.}

\item{pdf_file2}{Path to the PDF file 2.}

\item{output_file}{Path to the output Exccel (.xlsx) file.}

\item{n, }{carga horária do docende (padrão n=40)}
}
\description{
This function reads data from to PDF files and creates a Exccel (.xlsx)
document containing a table with the datas.
}
\examples{
\dontrun{
tabela_cad_xlsx(pdf_file1, pdf_file2, output_file, n=40)
}
}
