% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabtibble.R
\name{tabtibble}
\alias{tabtibble}
\title{Converte a tabela em texto em um dataframe}
\usage{
tabtibble(tabela)
}
\arguments{
\item{tabela}{texto com os dados do RADOC do docente}
}
\value{
um dataframe com os dados do texto
}
\description{
Converte a tabela em texto em um dataframe
}
\examples{
\dontrun{
tabtibble(tabela)
}
}
