% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topicos}
\alias{topicos}
\title{Tópicos}
\format{
\subsection{topicos}{

\describe{
\item{"I - ATIVIDADES DE ENSINO"}{}
\item{"I-1 Ensino Básico"}{}
\item{"I-1 Graduação"}{}
\item{"..."}{}
}
}
}
\usage{
topicos
}
\description{
Contém a lista de tópicos para a localização dos dados no SICAD+
}
\keyword{datasets}
