% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afastamentos.R
\name{afastamentos}
\alias{afastamentos}
\title{Afastamentos}
\usage{
afastamentos(file)
}
\arguments{
\item{file}{o RADOC do docente}
}
\value{
um dataframe com os dados do afastamento do docente
}
\description{
Afastamentos
}
\examples{
\dontrun{
afastamentos(file = 'radoc.pdf')
}
}
