% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docente.R
\name{afastamentos}
\alias{afastamentos}
\title{Função que extrai os dado de afastamentos do docente do RADOC em analise}
\usage{
afastamentos(pdf_file)
}
\arguments{
\item{pdf_file}{arquivo em pdf do RADOC do docente}
}
\value{
um dataframe com os dados do afastamento do docente
}
\description{
Função que extrai os dado de afastamentos do docente do RADOC em analise
}
\examples{
\dontrun{
afastamentos(pdf_file)
}
}
\keyword{internal}
