% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_cad.R
\name{coleta}
\alias{coleta}
\title{Extrai os dados a serem pontuados no RADOC}
\usage{
coleta(pdf_file)
}
\arguments{
\item{pdf_file}{arquivo em pdf do RADOC do docente}
}
\value{
um dataframe com os dados coletados no arquivo
}
\description{
Extrai os dados a serem pontuados no RADOC
}
\examples{
\dontrun{
coleta('data-raw/radoc1.pdf')
}
}
\keyword{internal}
