% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docente.R
\name{docente}
\alias{docente}
\title{Extrai os dados do Docente}
\usage{
docente(tabela)
}
\arguments{
\item{tabela}{texto com os dados do RADOC do docente}
}
\value{
um dataframe com os dados do docente
}
\description{
Extrai os dados do Docente
}
\examples{
\dontrun{
docente(tabela)
}
}
