% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesCheck.R
\name{filesCheck}
\alias{filesCheck}
\title{Função que faz 3 testes nos arquivos}
\usage{
filesCheck(pdf_file1, pdf_file2)
}
\arguments{
\item{pdf_file1, pdf_file2}{dois arquivos em pdf}
}
\value{
Uma mensagem com a avaliação feita
}
\description{
\enumerate{
\item Testa se o arquivo é um RADOC do SICAD+
\item Testa se os dois RADOCs são do memso docente
\item Testa se os dois RADOCs são de anos distintos
}
}
\examples{
\dontrun{
filesCheck(pdf_file1, pdf_file2)
}
}
