% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pontua.R
\name{pontua}
\alias{pontua}
\title{Pontua os dados do RADOC seguindo o Anexo II da Resolução Consuni 18/2017}
\usage{
pontua(df)
}
\arguments{
\item{df}{data frame com os dados do RADOC}
}
\value{
um dataframe com os dados e suas respectivas pontuações
}
\description{
Pontua os dados do RADOC seguindo o Anexo II da Resolução Consuni 18/2017
}
\examples{
\dontrun{
pontua(df)
}
}
