% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_cad.R
\name{tab_cad}
\alias{tab_cad}
\title{Extrai os dados e retorna um dataframe com a Tabela da CAD}
\usage{
tab_cad(pdf_file1, pdf_file2)
}
\arguments{
\item{pdf_file1, }{arquivo PDF com os dados do 1o RADOC}

\item{pdf_file2, }{arquivo PDF com os dados do 2o RADOC}
}
\value{
um dataframe com a Tabela CAD preenchida
}
\description{
Função que extrai os dados dos RADOC, usa o AnexoII para pontuar os dados e
retorna um dataframe com a Tabela da CAD preeenchida.
}
\details{
Essa função usa duas funções internas uma para coletar os dados e outra para
pontuar os dados.
}
\examples{
\dontrun{
tab_cad(pdf_file1, pdf_file2)
}
}
\keyword{internal}
