% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_cad.R
\name{tabela_cad}
\alias{tabela_cad}
\title{Retorna o dataframe pronto para impressão da Tabela Cad}
\usage{
tabela_cad(file1, file2, n = 40)
}
\arguments{
\item{file1, }{dataframe com os dados do 1o RADOC}

\item{file2, }{dataframe com os dados do 2o RADOC}

\item{n, }{carga horária do docende (padrão n=40)}
}
\value{
um dataframe com a Tabela CAD preenchida
}
\description{
Retorna o dataframe pronto para impressão da Tabela Cad
}
\examples{
\dontrun{
tabela_cad(df)
}

}
