% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_cad.R
\name{tabela_cad}
\alias{tabela_cad}
\title{Retorna o dataframe pronto para impressão da Tabela Cad}
\usage{
tabela_cad(file1, file2, n = 40)
}
\arguments{
\item{df}{dataframe com os dados do RADOC}
}
\value{
um dataframe com a Tabela CAD preenchida
}
\description{
Retorna o dataframe pronto para impressão da Tabela Cad
}
\examples{
\dontrun{
tabela_cad(df)
}

}
