% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_cad_tex.R
\name{tabela_cad_tex}
\alias{tabela_cad_tex}
\title{Create a LaTeX Document from PDF Data}
\usage{
tabela_cad_tex(pdf_file1, pdf_file2, output_file, n = 40)
}
\arguments{
\item{pdf_file1}{Path to the PDF file 1.}

\item{pdf_file2}{Path to the PDF file 2.}

\item{output_file}{Path to the output Latex (.tex) file.}

\item{n, }{carga horária do docende (padrão n=40)}
}
\description{
This function reads data from tow PDF files and creates a LaTeX (.tex)
document containing a table with the datas.
}
\examples{
\dontrun{
tabela_cad_tex(pdf_file1, pdf_file2, output_file)
}
}
