% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_cad.R
\name{tabela_cad_xlsx}
\alias{tabela_cad_xlsx}
\title{Cria a Tabela Cad (em Excel) a partir dos RADOCs}
\usage{
tabela_cad_xlsx(pdf_file1, pdf_file2, output_file = NULL, n = 40)
}
\arguments{
\item{pdf_file1}{o caminho para o arquivo contendo o 1o RADOC.}

\item{pdf_file2}{o caminho para o arquivo contendo o 2o RADOC.}

\item{output_file}{o caminho onde o arquivo de saída deve ser salvo.}

\item{n}{carga horária do docende (padrão n=40)}
}
\description{
Esta função lê os dados do RADOC em 2 arquivos PDF e retorna um
arquivo Excel (.xlsx) com a Tabela Cad preenchida.
}
\details{
Em caso de ausência do argumento \code{output_file} o arquivo será salvo na
pasta atual.
}
\examples{
\dontrun{
tabela_cad_xlsx(pdf_file1, pdf_file2, output_file, n)
}
}
